+++ kwimage/structs/polygon.py
@@ -1587,6 +1587,19 @@ class Polygon(_generic.Spatial, _PolyArrayBackend, _PolyWarpMixin, _ShapelyMixin
         xy = (shp_centroid.x, shp_centroid.y)
         return xy
 
+    def to_box(self):
+        """
+        Returns:
+            kwimage.Box
+        """
+        import kwimage
+        xys = self.data['exterior'].data
+        lt = xys.min(axis=0)
+        rb = xys.max(axis=0)
+        ltrb = np.hstack([lt, rb])
+        boxes = kwimage.Box.from_data(ltrb, 'ltrb')
+        return boxes
+
     def bounding_box(self):
         """
         Returns an axis-aligned bounding box for the segmentation
@@ -1883,6 +1896,10 @@ class Polygon(_generic.Spatial, _PolyArrayBackend, _PolyWarpMixin, _ShapelyMixin
             facecolor = color
         else:
             facecolor = kwimage.Color(facecolor, alpha=alpha)._forimage(image)
+
+        # TODO: consolidate logic
+        # _generic._handle_color_args_for(
+        #     color, alpha, border, fill, edgecolor, facecolor, image)
 
         if fill:
             if alpha is None or alpha == 1.0: